+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- fix imports in `impl_Attribute_for_Parse_and_ToTokens!`
 
 ## [0.10.2] - 2024-10-30
 ### Added