+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.10.3] - 2024-11-17
 ### Fixed
 - fix imports in `impl_Attribute_for_Parse_and_ToTokens!`
 ### Added
@@ -98,7 +100,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking Change**: Compile time verify if ident is given through helper
   trait
 
-[unreleased]: https://github.com/ModProg/attribute-derive/compare/v0.10.2...HEAD
+[unreleased]: https://github.com/ModProg/attribute-derive/compare/v0.10.3...HEAD
+[0.10.3]: https://github.com/ModProg/attribute-derive/compare/v0.10.2...v0.10.3
 [0.10.2]: https://github.com/ModProg/attribute-derive/compare/v0.10.1...v0.10.2
 [0.10.1]: https://github.com/ModProg/attribute-derive/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/ModProg/attribute-derive/compare/v0.9.2...v0.10.0